refactor: gate ai generated fake profile mapping
