+++ Pryan-Fire/projects/backend/anewluv-photo-moderation/src/photo_sweeper/model.py
@@ -13,7 +13,7 @@ from pathlib import Path
 from typing import Any
 
 from .moderation_contract import provider_instructions
-from .validators import default_profile_checks, normalize_model_result, safe_note_for_reason
+from .validators import default_profile_checks, normalize_minimax_description, normalize_model_result
 
 DEFAULT_MODEL_FIXTURE = Path(__file__).resolve().parent / "fixtures" / "mock_model_responses.json"
 DEFAULT_CODEX_MODEL_ROUTE = "Codex OAuth/OpenClaw gpt-5.5 + gpt-image-2 configured image route"
@@ -152,61 +152,6 @@ class MiniMaxCLIAdapter:
             return _manual_failure("api_failure_fallback", self.model, f"MiniMax CLI exited with status {proc.returncode}; manual review required.", fallback="manual_review")
 
         return normalize_minimax_description(_extract_text(proc.stdout), model=self.model)
-
-
-def normalize_minimax_description(description: str, *, model: str = DEFAULT_MINIMAX_MODEL) -> dict:
-    text = description.lower()
-    checks = default_profile_checks(needs_human_review=True)
-    unsafe_categories: list[str] = []
-    verdict = "review"
-    reason_code = "manual_review_needed"
-    confidence = 0.55
-
-    if any(term in text for term in ["porn", "pornographic", "explicit", "genital", "sexual act", "intercourse", "masturbat", "nude", "nudity"]):
-        verdict = "reject_recommendation"
-        reason_code = "sexual_content"
-        confidence = 0.88
-        unsafe_categories = ["sexual_content"]
-    elif any(term in text for term in ["ai-generated", "ai generated", "generated image", "illustration", "cartoon", "anime", "drawing", "rendered"]):
-        reason_code = "ai_generated_image"
-        confidence = 0.78
-        unsafe_categories = ["ai_generated_image"]
-        checks["is_ai_generated"] = True
-    elif any(term in text for term in ["screenshot", "meme", "advertisement", "phone number", "email", "social media handle", "qr code"]):
-        reason_code = "contact_info_or_ad"
-        confidence = 0.74
-        unsafe_categories = ["contact_info_or_ad"]
-        checks["has_contact_info"] = True
-        checks["is_meme_or_screenshot"] = True
-    elif any(term in text for term in ["blurry", "dark", "unclear", "obscured", "low quality", "not visible"]):
-        reason_code = "low_quality_or_unusable"
-        confidence = 0.68
-        unsafe_categories = ["low_quality_or_unusable"]
-        checks["is_blank_or_unusable"] = True
-    elif any(term in text for term in ["person", "people", "man", "woman", "face", "portrait", "selfie", "individual", "subject"]):
-        verdict = "approve_recommendation"
-        reason_code = "clean_profile_style"
-        confidence = 0.82
-        checks["is_profile_style_photo"] = True
-        checks["needs_human_review"] = False
-    else:
-        reason_code = "not_a_profile_photo"
-        confidence = 0.62
-        unsafe_categories = ["not_a_profile_photo"]
-
-    return normalize_model_result(
-        {
-            "verdict": verdict,
-            "confidence": confidence,
-            "reason_code": reason_code,
-            "note": safe_note_for_reason(reason_code),
-            "unsafe_categories": unsafe_categories,
-            "vision_model_used": f"{model} + parser",
-            "fallback_model": None,
-            "app_profile_photo_checks": checks,
-        },
-        default_model=f"{model} + parser",
-    )
 
 
 def _provider_instructions() -> str:
